v_learn_backend/requirements.txt
Flask==2.3.3
flask-jwt-extended==4.4.4
pymongo[srv]==4.5.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
pytest==7.4.0
gunicorn==21.2.0
